Blame actions in the sidebar of each file
Add "Actions" at the top of the sidebar, including Log, Blame, Diff and Raw.
Each one can be overridden in dxr.config. hg, git and bzr are supported for
determining the current revision, alternatively "revision=" can contain a
custom command. A "Notice:" is printed once if the repository is not hg.

Fixes: https://bugzilla.mozilla.org/show_bug.cgi?id=701411