Add FTS4 code-friendlier tokenizer
This tokenizer will provide tokens for both the joined and
split forms of CamelCase and underscore_separated strings,
enabling searches of full function/class names, or parts of
these. This results in more coherent results from code searches.

The module loading is a bit convoluted, and could be simplified
in the future if python sqlite bindings enabled extension loading.
It currently requires a module function to be called through ctypes,
and a call to a custom sqlite function to enable the tokenizer
for the calling handle.