clang: canonicalize class declarations straight away
The clang plugin used to generate distinct "type" entries in the
csv files for declarations like:

class FooBar;

And the indexer plugin would normalize those to point to the
real definition of the class. This step is now done directly
in the clang plugin, as it's less guess work and the data can
make it directly to the database.