dxr: Fix redundant results
Quoting a single search term in the FTS query happened to
confuse the tokenizer, having it return multiple results
per line, so only add quotes if the term contains '-',
which is part of the FTS syntax.

Also, ensure that a line is returned once even if it contains
the search term multiple times.