Add profiling ability for SQL queries.
Appending "&explain=true" to the end of the URL of any query will run
the query as normal but not return the results to the browser.
Instead it will report the results of "EXPLAIN QUERY PLAN" for every
SQL statement that is run along with the time that each statement took.

This can be useful in diagnosing slow queries and figuring out how to
speed them up.