dxr-build: add ignore patterns involving the whole path
This adds support for ignore_patterns that take the whole path into
account.  Any ignore patterns that start with "/" will be matched against
the whole path relative to the root of the source tree, instead of just the
leaf name of the item.

Note that this doesn't do the right thing - it will happily match sub-
directories with a single "*".