Move some files around in preparation for turning the Python parts into a legitimate package. (Tests should fail.)
The dxr folder is going to form the core of an installable package. search.py, query.py, and utils.py never change; they just get copied on build. Thus, those can live in the installed package, and the build can just be a config file and .htaccess. To serve, run dxr-serve.py --config=/my/build/dir. We can explain that in the last 2 lines output from dxr-build, much as we currently do with test-server. For WSGI, we can pass in the path to the config file in an env var.

Because this reduces repetition, it makes dev easier: we can just edit the single, installed (with python setup.py develop) version of the server, and git will pick it up—no more having to edit an untracked build artifact. More standard installation and project layout probably won't hurt our community adoption, either.

Note that the package won't encompass the entire project, as there's no nice way to make setup.py build trilite and the clang plugins (that I know of).