[bug 826482] Fix double slashes appearing in URLs.
These aren't tolerated by the test server anymore.

While I was at it, rewrote build_folder so it's a little (but not much) easier to follow. Took advantage of the filesizeformat Jinja filter so our file sizes are a little nicer (decimals, like the previous author wanted) and our code for dealing with them a little shorter.