Stop generating .htaccess files.
Those are really, really clumsy (impossible?) to get working with WSGI. Also, they're not great for performance. If you're going to put a production deployment, it's not a bad assumption that you also control the web server and can put in global config. If you're just booting up DXR to play around, you can use dxr-serve.