Specify dependencies in setup.py. Flatten out requirements. Install python libs from source.
Our Vagrant VM is Ubuntu, but we'll deploy on RHEL, and they'll probably have different sets of packaged Python libs. To avoid having an explosion of intertwingled requirements files, we'll just install everything from pip for now. Our thought for production deployment is to have the admin node build the compiled (and other) libs, since that box is architecturally and shared-lib-ly identical to the webheads.

Also, get rid of python-wsgi-intercept, which we don't need.