Add constructor call information to clang plugin
This is just a slightly modified copy of the regular function/method
call code. Unfortunately there's no place to click on the constructor
invocation to find the definition, but on clicking on the definition
you can find callers and references. I've added a creation of a
BitField object HelloWorld/main.c as an example (click on the BitField
constructor, and main in main.c can be found as a caller).