Make fully qualified member names trigger direct results. Fix bug 766122.
For some unknown reason, the % was at the front rather than at the back, where it's needed to eat the () or (params) which we jam into the qualname field along with the function name. (Type-signature overloading and all that jazz)

Also add a test for the other qualified-name direct-result case: types.