Stop using suffix matching for direct-result qualified-type queries.
This ought to make them much faster, as they'll be able to use the index (assuming the collation is case insensitive, which I'm not sure about). They're covered by the test I added in the previous commit. I've jazzed it up to check case-insensitivity, which may or may not be desired. For now, I'm playing it safe.