Replace ad hoc reimplementation of multiprocess pooling for HTML build phase with concurrent.futures.
The code we're deleting has been the source of multiple bugs. Also, we no long have to cleverly go dig up the right interpreter and the dxr-worker.py script to run with it.

Log files are differently named now, with the bonus that you don't have to go through as many layers of indirection to figure out which files was handled by which worker. Also, overall progress indication is given: we know upfront how many workers there will be.

There's no more command-line dxr-worker.py to run, but there was never any reason to run that manually anyway.