Don't retry fetching results on failure, 500 or otherwise. Fix bug 848480.
I'm not sure why we were being so failure-tolerant in the JS before; perhaps the CGI failed or timed out a lot. With WSGI, we no longer have timeouts to worry about, and, if we have other failures, we should solve the root problem.

If there is an error, the user can "retry" just by changing the search query or hitting Return, both of which should be natural reactions.