Rewrite build_lines so we no longer emit misbalanced tags and to support overlapping and line-spanning regions.
Add support for overlapping regions (not refs, because that would be a UI problem). Preserve the semantics of interleaved refs by translating to an equivalent balanced series of tags. This lets plugins be less coupled with each other: deployers no longer have to worry about which plugins get along with each other: even if we use different tokenizers for a language (as we do with clang and pygments), build_lines will reconcile their token boundaries.

It's also got test coverage now.