Rip out the remainder of custom template support.
The instructions were hopelessly out of date, we had already stripped out the other example templates, and we've genericized the included one to suit non-Mozilla deployments.

The motivation to do this now is that we need to share some templates between the client and server side, and figuring out how to support that in a custom template--which nobody's using anyway--makes it more complicated.

Also remove the foot_text option. If we want something like this, it should be HTML anyway so you can have links.

In layout.html, fix the footer block so it doesn't leave an unbalanced div tag, and fix whitespace so we're not deceived into doing that again.