Make entire rows of the folder-contents tables clickable.
...with no JS (so we get to see link addresses on mouseover, command-click to open in a new window, etc.), non-linked padding between columns, or other horrors. Please tell me there are no other horrors.

Essentially, made the <a> tags display-block, and moved the padding from the table cells to the <a> tags themselves. Then I had to move the icon positioning around to compensate. The triple repeat of the links in the markup is unfortunate, but I thought it better than copying it around with JS. It should pretty much zip out.

Also shifted the folder icon down a tad so all the icons are level with the text baseline.