Lighten greys to match mockup. Standardize hover color across various tables and menus. Factor up colors as much as possible so we don't destandardize accidentally. Ref bug 959863.
I intend that whitening the Switch Tree menu and the odd rows of the folder-contents table will provide a bar of white to delimit the search panel from what comes below. But I think we need to increase the left margin of what's below to make that work. We'll see if it works out.