Upgrade Vagrant box to saucy. Lose puppet. Upgrade LLVM and clang.
Motivation: this gives us a new enough npm to run all our npm stuff.

As a bonus, we get packaged versions of LLVM and clang and can stop yanking down binaries from llvm.org all the time, so we delete the install_llvm script.

Swapped out puppet for a simple shell script because puppet was stuck in a CPU-pegging loop during "vagrant up". strace said it was just calling getdent() over and over again. I've had enough of puppet and its overcomplication and unsuitable failure modes for our purposes. [I later found out that it was out-of-date guest additions that were causing the loop, but I've been meaning to oust puppet for some time.]

While I was at it, upgraded Vagrant box to LLVM and clang 3.3.