Add a spinner to indicate when a search is in flight. Fix bug 966003.
Hoping Retina displays use the extra data in the extra-big image.

Attaching it as a background image to the Filters menu saves us having to (1) guess how wide the Filters menu is and pad out an absolutely positioned spinner element to not crash into it or (2) jam the Filters menu and the spinner inside a common parent element, which struck me as an unnecessary pollution of the semanticness of the markup.