Remove spurious padding around anchors in file views' code. Fix bug 969887. Keep code from jumping to the right when a line number is clicked. Fix bug 969884.
Stop classing file views as result snippets. Refactor CSS to compensate.

Remove the gold border and "pull-out" effect for now. Not thrilled with the resulting cosmetics, but we can come back later and think of something.

This will need a re-index to take effect, but it looks good enough with the old markup to not bump the format version.