Do nav panel hiding (and showing) server-side. Speed up the transitions as well. Refs bug 968020. Bump format version.
There's no reason to bring JS into this. Now we shouldn't have to wait for the whole, long page of source code to display before we can see the nav.

Also remove a hard-coded-true test and an unused block from the file template and a useless test for 0 length.

Bump the format version. Otherwise, the discloser on the nav pane is always in the wrong state.