Fix the warnings tests.
What we're doing now may not be entirely equivalent to what we were doing for extents. We're now getting the spelling location of a spelling location, since we're now calling locationToString() (which now gets the spelling location) on the result of getWarningExtrentLocation() (which gets it itself). I'm not sure if that takes things through another round of expansion, for instance. But there's a great deal of cleanup to be done in the clang plugin in any case, so expedience wins the day here.