Stop supporting the `dxrroot` config option.
It makes more sense to remove this option than to document it. Its main purpose is to point to a custom templates directory, and we previously removed support for custom templates. It can also be used, indirectly, to point to a custom plugins dir, but that's also about to be obsolete, with the upcoming new plugin system.