Do urllink in the new plugin API. Correct some of the problems it revealed.
Rip out Unicode decodes; we'll be passing Unicode in, so we don't need to decode.

I don't actually like how this is turning out. The too-many-objects smell turned out to be telling the truth: you end up paying for what you don't eat when writing a simple plugin. Committing this before trying a more fractal approach.