Pass the TreeConfig to file skimmers.
Otherwise, they have no way to get at it; they have no parental TreeToIndex to pass them things.

Pass the tree to file indexers as well. This is mostly for symmetry, but it might remove the need for a TreeToIndex in cases where it was used merely to extract config data.