Make test harness ES-savvy.
As a bonus, it now calls build_instance() directly rather than shelling out to dxr-build.

Tell all the test config files to name their ES indices starting with "dxr_test_". We'll remove this and a lof ot other boilerplate when we refactor the config system.

Replace ES_INDEX with ES_ALIASES in the request-time config file. We don't need to know the index names, but we do need to know one alias per tree.