Fix the FILE/LINE superset problem with documentation for now. Fix bug 1058945. Close #331.
We can change the API to make it more automatic if it turns out to be a burden for plugin. In that case, most likely just move the folder, name, size, modified, and is_folder fields (the ones that aren't touched by queries) out of the core plugin and into build.py itself, then have something deep_update() LINE mappings with their corresponding FILE ones.