Allow filters to punt by returning empty. Do that in TextFilter for short terms. Ref bug 1060111.
This allows us to ignore shorter-than-trigram textual query terms without saying anything, just as the production DXR does. (There's no need for us to call attention to our deficiencies; I fear that would noise up the feedback we get when we stage the ES branch. A real solution will be reached in the Query Language Refresh. Most likely, we will merge adjacent textual terms into phrases, at least as one of our search strategies.)

As in production, highlight shorter-than-trigram terms even if we're not able to search on them.

Rename BadQuery to BadTerm; we're not giving up on the whole query if one term won't run.