Skip the tests whose functionality isn't implemented yet: mostly clang ones.
We now have the ES groundwork laid and are working toward passing the test suite. It's distracting to have so many spurious failures. Now, if something fails, it's a problem. There are 3 of these at present.

Unskip tests as you implement what they're testing.

Remove test_filters altogether, since the one thing it tests is dead now and will be deleted when the whole filters module goes away.