Reorganize plugin base and util code to dodge circular imports and improve comprehensibility.
Now nothing is in the plugins/ folder except for actual plugins. "Utility" functions (which make up an awful category, equating to "other") are categorized as indexer-related or filter-related and shunted to appropriate top-level dxr modules accordingly.

Hoping that this will also untangle the import graph enough that we can start declaring clang's Plugin explicitly and not have to repeat ourselves every time we add a new clang filter.