Audit compiler plugin for weird location handling.
The bits that spit out the extents never made any sense to me. Many tunneled into macro definitions (spelling locations) and obeyed #line directives (presumed locations) when they shouldn't have. I suspect this is a major (or the only) cause of the completely bogus extents (for which I added warnings) that are emitted when analyzing mozilla-central.

24 test failures. Next, we'll tamp the test failures down to 0, and then we'll run against mozilla-central again and see if we improved things.

Also changed a name, "decl" to "def", which was wrong.