Make computation of class hierarchy and method overriding 300,000 times faster for bad cases. Fixes bug 1122685.
There were duplicate values in the graph. Apparently, clang sometimes spits out "B is a subclass of A" etc. more than once. Recursing down the Cartesian product of those was killing us. Now we eliminate the dupes.

Indexed moz-central (it was already built and the CSV files in place) in 108 minutes. That's workable, and we can probably improve it more later.