Upgrade to ES 1.4(.4).
The motivation is to make the mysterious hang-on-final-refresh go away for large indices.

Looks like ES is no longer willing to implicitly pull the first item out of array-type fields (like "content") anymore. That's actually good; I'm surprised that ever worked. But also, .value no longer returns a defined value, in conflict with the docs. Plain old [0] seems to do it.