[python-plugin] Fix the import resolution algorithm for base class names.
The plugin attempts to track names imported into modules in
TreeAnalysis.names[], which maps
  * the local name of an import in the module’s scope
  * to its original name and the module it was imported from.

For example, if we had the following in package/__init__.py:
    from package.foo import bar
    import package.foo.baz
...the following mappings would be stored:
    (package, bar) —is-really—> (package.foo, bar)
    (package, package.foo.baz) —is-really—> (GLOBAL, package.foo.baz)

Later, when processing a reference to X from `package`'s scope, we try
two things:
1) If the reference is something we have in our list of known imported
names:
  a) a module - the `package.foo.baz` case above
  b) a class or module imported from a module (`bar`)
...we take the original name and module it was imported from and check
again.
2) If the reference is of form `bar.Class`, we check what `bar` refers
to => (package.foo, bar.Class)

Before this change, a single ‘dotted’ string like `package.subpackage.module.var` was used to identify both the module and
the local name inside the module. This broke on dxr code itself:
  # in dxr/cli/__init__.py
  from dxr.cli.deploy import deploy

Creates this mapping:
  'dxr.cli.deploy' -> 'dxr.cli.deploy.deploy'
.. which is circular, since it defines d.c.d in terms of itself. So when
we attempted to resolve a class like `dxr.cli.deploy.object` we'd get:
  `dxr.cli.deploy.object` ->
  `dxr.cli.deploy` + `object` ->
  `dxr.cli.deploy.deploy` + `object`, resulting in an endless loop.

The solution is to use tuples (absolute_module_name,
local_name_inside_the_module), and never resolve the
absolute_module_name part.

This does mean that this analysis will not work in cases where we import
from a non-absolute name, like `submodule` in example below:
  from package import submodule
  from submodule import MyClass as Klass
  class ThirdDerivedFromSubmodule(Klass):
      pass
...but that didn't work before either.