Handle permalink target for VCS root different from tree folder.
Originally we construct the permalink path from the relative path
to the vcs; however when app.py fetches the revision it expects
a path relative to the tree’s source folder. When the vcs root
is not coincident with the tree’s source folder then we have a
mismatch. Fix the problem by always storing paths relative to
tree source folder, and using relpath to get a vcs relative path
when we query for the revision contents.