Store type-stripped qualnames for functions as well, so we can search for Scoped::function rather than needing to say Scoped::function(int, char). Fixes bug 1171242.
We store them as a second item in function needles' qualname properties, which becomes a list. This causes the right thing to Just Happen in QualifiedNameFilterBase, which needed only a slight adjustment to tolerate qualnames that are lists.

Added an arg to some cases in test_functions, because we weren't testing that anywhere, and I don't trust future versions of clang to keep the "()" on the end of argless function qualnames.

Also, replace the keys kwarg of needles() with a simple include_qualname boolean. The desired keys don't seem to be expanding, and this way we don't have to repeat the names of the keys in the callers.