Add same-symbol highlighting. Fixes bug 1123877.
This is a quick and dirty pass to bench against. This still returns tuples from refs(), but I'm starting to see that we'll need to expand the return values in the future, so we should probably switch to class instances or something. However, I'm concerned about performance, since tuples construct about 4x faster than even slotted instances. So this implementation is a mostly a version to bench against, to see if that's even a significant part of the build time.