Catch UnicodeDecodeError when doing the initial analysis on Python files.
If a file is encoded in some other encoding and has a sequence of
bytes that prevents it from being interpreted as UTF-8, it'll get
slurped up as a bytestring instead of a unicode string.  And in that
case, it'll probably hit the ast_parse utility function and throw this
error.