Add a --uid arg to the dev Dockerfile for those intrepid souls on Linux and calling docker directly.
docker-compose does not yet support the --build-arg arg: https://github.com/docker/compose/issues/2111. When it does, we can set this from the makefile, solve the problem of non-matching UIDs for Linux devs, and let us get rid of the chmod in Travis.

Also combine 2 ENV commands to save a snapshot.