Fix a segfault during clang CSV ingestion. Fixes bug 1246700.
It was triggered by the compiler plugin emitting cycles and then manifested as a segfault rather than a RuntimeError because we were nesting itertools.chain() so deeply.

For now (since moz-central is a week behind), I've simply added detection of cycles to _walk_graph()--and moved detection of dupes up into it, since it happens as a natural matter of course. Ehsan has expressed interest in moving the full inheritance computation to the compiler plugin in the longer term. Got through an entire moz-central index with workers=0. The chain() calls didn't survive, since they no longer fit with the imperative style needed to keep track of the seen set. It would be interesting to see if this is faster (since it prunes the graph traversal earlier) or slower (since it spends more time in the interpreter) than the old impl.