Only push browser history for user-typed queries.
Currently we push the result of every doQuery to the browser history, which
leads to history squashing in certain cases.  For example if we load a page
containing a search query (either from reloading a page or browsing through
history), then the page load gets a history entry, and then we do a doQuery for
the url query and that produces an identical history entry which deletes any
forward history.  It can also require (depending on where in history you're at)
two quick backs (before the next doQuery completes on the first back!) to
escape.

We now only push a new history item when it's the result of user typing in the
query box.