Change some ok_'s to eq_'s in testing.
ok_ is being used in some places where eq_ was intended.

Also, change test_some_bytes_visible in test_binary_files.py to drop the test
checking that we 404 on a /source/ page for binary files - we're now supporting
that scenario (and haven't been 404ing in any case, at least not in the recent
past).