Use binaryornot and chardet (as a backup) for text detection.
binaryornot is replacing our old simple is_text, and chardet is being used
as a backup when we think a file is text but we're unable to decode it using
the configured source_encoding.  It might be nice if we could just drop
source_encoding all together, but chardet is quite slow - running chardet over
every is_text file in mozilla-central leads to indexing times nearly twice
their current length.