Cut off python name inheritance cycles.  Fixes bug 1264627.
The bug is caused by a class definition in
mozilla-central/testing/web-platform/tests/tools/py/py/_log/warning.py of
'class DeprecationWarning(DeprecationWarning):', where the base
DeprecationWarning is a built-in which is never explicitly imported.  We don't
distinguish between the new DeprecationWarning name and the built-in name, so
it looks to us like we have an inheritance cycle.  Similar bugs can occur
with longer cycles.

The fix here cuts those off so that we don't crash on indexing when we exceed
the allowed recursion depth.