Remove 'no_from'; break back button redirect-to-single-result loop.
"Jump to definition" is a link to a redirect query which creates its own history
entry, so you click "Jump to definition", it loads the query page, and the query
runs and redirects to a single result.  With the previous behavior, if you then
clicked back you would go back to a redirect query which then just redirects you
back to the single result again.  Now when we load a page with a redirect query,
we do the query but then replace the url with its non-redirect version so that
when you go back to the query page you stay there.