Change directions. Embrace paths as bags of bytes. Fixes bug 1290945.
This is how Linux handles them, so we may encounted undecodeable names as we traverse the FS. Convert to unicode only at the last minute, for ES and for display on the web. This may cause dirs and files that are named the same except for undecodeable bytes to collide: dirs may merge, or files may be shadowed. This is still an improvement over indexing crashing.

Add an undecodeable folder name and file name to the non-ASCII test.

Clarify that plugins should expect incoming paths to be bytestrings.