Fit switch tree menu
The expanded switch tree menu wasn’t fitting inside viewport. Gave it
max-height limits (70% of viewport for browsers supporting ‘vh’, and
250px for those that don’t), and made it scrollable (if it exceeds
max-height).

Another reason I had to make this change was due to putting the tree in
a fixed nav bar. Setting position to ‘fixed’ was causing the expanded
menu to be clipped, regardless of overflow setting.