Refactored templates
Moved all of the nav_bar block to layout.html and removed it from both
file.html and folder.html.

Note: In the previous blocks the if-condition inside the loop was a
little different for files and folders (folders had an extra “and
loop.length != 1”, but it seemed safe enough to include the extra
condition in files too (I don't think the loop.length will ever be 1).