Disabmiguate static functions that share the same name but appear in different files.
Before this change if you had a function like "static void foo()" in
one file and another, different function also called "static void foo()"
in a different file, DXR would not be able to entirely distinguish
between the two.  For instance, if you clicked on one of the
definitions and selected "Find references" then the list that was
produced would include references not only to the selected function but
also to the other function.