Fix user-defined literals.
C++ introduced a new thing called "user-defined literals".  These have double
quotes in their name.  This was causing problems when writing out the
temporary .csv files.  If there is a double quote in a CSV field then the
quote should be escaped by doubling it.  There was already code to do this
but it was only enabled for certain fields.  Identifiers were assumed to
never contain quotes and thus not need escaping.

Since using the proper quoting algorithm will always produce the correct
result for any type of field, turn it on unconditionally for all fields.