Fix missing "Jump to definition" link on function declarations.
I believe I broke this in 94882fe.  There is now, in some circumstances,
the possibility of having multiple decldefs for a single function.  Only one
will have a defloc and the rest will be without it.

When we're building the page we sort them (in essentially random order) and
pick the first one.  If that one happens to be one without a defloc then the
"Jump to definition" menu item will not be shown.

The fix here makes sure that if we have a decldef with a defloc then it is
chosen in preference to others without a defloc.  If all the decldefs that
we have are without deflocs then the one that we pick is essentially chosen at
random like before.  But in this case all the decldefs should be identical so
this should not cause any issues.